refactor(games): extract trivia service
