+++ bot/features/games/__init__.py
@@ -1,0 +1,13 @@
+"""Game feature services."""
+
+from .trivia_service import (
+    JsonTriviaLoader,
+    TriviaLoader,
+    TriviaService,
+)
+
+__all__ = [
+    "JsonTriviaLoader",
+    "TriviaLoader",
+    "TriviaService",
+]
